feat : 이름을 담는 Name 클래스 생성 이름에 대한 검증 로직을 Name 내부에서 진행
