+++ src/main/java/blackjack/domain/user/Player.java
@@ -6,24 +6,14 @@ import blackjack.domain.card.Number;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 public class Player {
-    private static final Pattern PATTERN = Pattern.compile("^[가-힣a-zA-Z]*$");
-    private static final String PLAYER_WRONG_NAME_EXCEPTION_MESSAGE = "이름을 잘못 입력하였습니다. (입력값 : %s)";
     private final List<Card> cards = new ArrayList<>();
-    private final String name;
+    private final Name name;
 
     public Player(String name) {
-        validate(name);
-        this.name = name;
-    }
-
-    private void validate(String name) {
-        if (!PATTERN.matcher(name).matches()) {
-            throw new IllegalArgumentException(String.format(PLAYER_WRONG_NAME_EXCEPTION_MESSAGE, name));
-        }
+        this.name = new Name(name);
     }
 
     public List<Card> getCards() {
@@ -77,7 +67,7 @@ public class Player {
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public List<String> getCardsStatus() {
